+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* correct typo: esrifieldTypeInteger => esriFieldTypeInteger
+
 ## [2.3.1] - 2017-06-15
 ### Fixed
 * Serialize date fields as UNIX timestamps